Reuse stripe error message mapping
